Create error content component
